refactor(input): separate shake logic from CustomInput into standalone ShakeInput component
